feat: add useURL in shared
